+++ examples/with-sveltekit/src/routes/__layout.svelte
@@ -1,5 +1,5 @@
 <script context="module">
-  import { setup } from '@twind/sveltekit'
+  import { setup } from '@twind/with-sveltekit'
   import config from '../twind.config'
 
   setup(config)