+++ packages/sveltekit/hooks.d.ts
@@ -1,3 +1,0 @@
-// ONLY FOR TYPESCRIPT TO FIND @twind/sveltekit/hooks
-
-export * from './src/hooks'