+++ packages/with-sveltekit/hooks.d.ts
@@ -1,0 +1,3 @@
+// ONLY FOR TYPESCRIPT TO FIND @twind/with-sveltekit/hooks
+
+export * from './src/hooks'